+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to Semantic Versioning.
 
 * File drag and drop events can now be detected via `Event::FileDropped`.
 * The clipboard can now be manipulated via `input::get_clipboard_text` and `input::set_clipboard_text`.
+* `input::get_key_modifier_down` and `input::get_key_modifier_up` have been added, allowing for code handling the control, alt and shift keys to be made more compact.
 
 ## [0.3.4] - 2020-04-12